# april27finale-proper-data-addition/april27finale-proper-data-addition/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: QuickBooks-style business platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductInput"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductInput"
      responses:
        "200":
          description: Product updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Product deleted
...
  /invoices:
    get:
      operationId: listInvoices
      tags: [invoices]
      summary: List all invoices
      responses:
        "200":
          description: List of invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invoice"
    post:
      operationId: createInvoice
      tags: [invoices]
      summary: Create an invoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvoiceInput"
      responses:
        "201":
          description: Invoice created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /invoices/{id}:
    get:
      operationId: getInvoice
      tags: [invoices]
      summary: Get an invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
    patch:
      operationId: updateInvoice
      tags: [invoices]
      summary: Update an invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInvoiceInput"
      responses:
        "200":
          description: Invoice updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
    delete:
      operationId: deleteInvoice
      tags: [invoices]
      summary: Delete an invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Invoice deleted
  /invoices/{id}/pay:
    post:
      operationId: payInvoice
      tags: [invoices]
      summary: Mark invoice as paid
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PayInvoiceInput"
      responses:
        "200":
          description: Invoice marked as paid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
...
components:
  schemas:
...
    LineItem:
      type: object
      properties:
        productId:
          type: integer
          nullable: true
        description:
          type: string
        quantity:
          type: number
        unitPrice:
          type: number
        taxPercent:
          type: number
          default: 0
        discountPercent:
          type: number
          default: 0
      required:
        - description
        - quantity
        - unitPrice
...
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        sku:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        salePrice:
          type: number
        costPrice:
          type: number
        taxPercent:
          type: number
          default: 0
        discountPercent:
          type: number
          default: 0
        discountAmount:
          type: number
        minOrderQty:
          type: number
        preferredVendorId:
          type: integer
          nullable: true
        isInventoryItem:
          type: boolean
        unit:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - salePrice
        - costPrice
        - discountAmount
        - minOrderQty
        - isInventoryItem
        - createdAt
        - updatedAt
    CreateProductInput:
      type: object
      properties:
        name:
          type: string
        sku:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        salePrice:
          type: number
        costPrice:
          type: number
        taxPercent:
          type: number
          default: 0
        discountPercent:
          type: number
          default: 0
        discountAmount:
          type: number
        minOrderQty:
          type: number
        preferredVendorId:
          type: integer
          nullable: true
        isInventoryItem:
          type: boolean
        unit:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      required:
        - name
        - salePrice
        - costPrice
    UpdateProductInput:
      type: object
      properties:
        name:
          type: string
        sku:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        salePrice:
          type: number
        costPrice:
          type: number
        taxPercent:
          type: number
        discountPercent:
          type: number
        discountAmount:
          type: number
        minOrderQty:
          type: number
        preferredVendorId:
          type: integer
          nullable: true
        isInventoryItem:
          type: boolean
        unit:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
...
    Invoice:
      type: object
      properties:
        id:
          type: integer
        customerId:
          type: integer
        customerName:
          type: string
        estimateId:
          type: integer
          nullable: true
        quoteId:
          type: integer
          nullable: true
        trackingNumber:
          type: string
          nullable: true
        salesLead:
          type: string
          nullable: true
        status:
          type: string
          enum: [draft, sent, pending, paid, overdue, cancelled]
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/LineItem"
        subtotal:
          type: number
        taxTotal:
          type: number
        discountTotal:
          type: number
        total:
          type: number
        dueDate:
          type: string
          format: date-time
          nullable: true
        paymentMethod:
          type: string
          enum: [stripe, bank_transfer, check, cash]
          nullable: true
        paymentNote:
          type: string
          nullable: true
        paidAt:
          type: string
          format: date-time
          nullable: true
        isQuickInvoice:
          type: boolean
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - customerId
        - customerName
        - status
        - lineItems
        - subtotal
        - taxTotal
        - discountTotal
        - total
        - isQuickInvoice
        - createdAt
        - updatedAt
    CreateInvoiceInput:
      type: object
      properties:
        customerId:
          type: integer
        estimateId:
          type: integer
          nullable: true
        quoteId:
          type: integer
          nullable: true
        trackingNumber:
          type: string
          nullable: true
        salesLead:
          type: string
          nullable: true
        status:
          type: string
          enum: [draft, sent, pending, paid, overdue, cancelled]
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/LineItem"
        dueDate:
          type: string
          format: date-time
          nullable: true
        isQuickInvoice:
          type: boolean
        notes:
          type: string
          nullable: true
      required:
        - customerId
        - lineItems
    UpdateInvoiceInput:
      type: object
      properties:
        customerId:
          type: integer
        status:
          type: string
          enum: [draft, sent, pending, paid, overdue, cancelled]
        trackingNumber:
          type: string
          nullable: true
        salesLead:
          type: string
          nullable: true
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/LineItem"
        dueDate:
          type: string
          format: date-time
          nullable: true
        notes:
          type: string
          nullable: true
        internalNote:
          type: string
          nullable: true
    PayInvoiceInput:
      type: object
      properties:
        paymentMethod:
          type: string
          enum: [stripe, bank_transfer, check, cash]
        paymentNote:
          type: string
          nullable: true
      required:
        - paymentMethod
...